requirements.txt
tensorflow
keras
numpy==1.26.4
pandas
matplotlib
tqdm
pillow